start victory checking
