+++ connect_four/bot_player.rb
@@ -1,0 +1,7 @@
+# require './player.rb'
+
+class BotPlayer < Player
+  def choose_column
+    (rand(7) + 1)
+  end 
+end
