+++ connect_four/execute.rb
@@ -1,0 +1,9 @@
+require './connect_four.rb'
+require './board.rb'
+require './win_judge.rb'
+require './player.rb'
+require './human_player.rb'
+require './bot_player.rb'
+
+game = ConnectFour.new
+game.play
